chore: move source to `src` and `ex`
