+++ Makefile
@@ -1,7 +1,7 @@
 CC = gcc
 
 all:
-	$(CC) -o vector.ex vector.ex.c
+	$(CC) -o vector.ex ex/vector.ex.c
 
 clean:
 	rm -f vector.ex
