+++ ex/vector.ex.c
@@ -1,4 +1,4 @@
-#include "vector.h"
+#include "../src/vector.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
